Fungus tree is now offered by Myssil after finishing a quest for her
git-svn-id: http://svn.net-core.org/repos/t-engine4@5090 51575b47-30f0-44d4-a5cc-537603b46e54

+++ src/map.h
@@ -34,7 +34,7 @@ struct s_map_object {
 	int cur_x, cur_y;
 	float dx, dy, scale;
 	float animdx, animdy;
-	int dw, dh;
+	float dw, dh;
 	float tint_r;
 	float tint_g;
 	float tint_b;
